+++ YIM/.idea/misc.xml
@@ -7,4 +7,11 @@
   <component name="ProjectType">
     <option name="id" value="Android" />
   </component>
+  <component name="VisualizationToolProject">
+    <option name="state">
+      <ProjectState>
+        <option name="scale" value="0.2331896551724138" />
+      </ProjectState>
+    </option>
+  </component>
 </project>
