+++ YIM/app/src/main/java/com/example/yim/vista/vista/VerEjercicios.java
@@ -1,4 +1,4 @@
-package com.example.yim.vista.vistas;
+package com.example.yim.vista.vista;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,11 +6,11 @@ import android.os.Bundle;
 
 import com.example.yim.R;
 
-public class PreguntasRegistro extends AppCompatActivity {
+public class VerEjercicios extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_preguntas_registro);
+        setContentView(R.layout.activity_ver_ejercicios);
     }
 }
