--- YIM/app/src/main/java/com/example/yim/vista/vistas/MainActivity.java
@@ -1,16 +1,0 @@
-package com.example.yim.vista.vistas;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-import com.example.yim.R;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-    }
-}
