+++ YIM/app/src/main/res/drawable/_style_borde_gris_arriba.xml
@@ -1,0 +1,10 @@
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:bottom="-2dp" android:left="-2dp" android:right="-2dp">
+        <shape android:shape="rectangle">
+            <solid android:color="@android:color/transparent" />
+            <stroke
+                android:width="2dp"
+                android:color="@color/gris" />
+        </shape>
+    </item>
+</layer-list>
