+++ YIM/app/src/main/res/drawable/_style_rectangulo_fondo_clarito__borde_blanco.xml
