+++ YIM/app/src/main/res/drawable/_style_rectangulo_gris_clarito__borde_negro_5.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <stroke
+        android:width="2dp"
+        android:color="@color/negro" />
+    <padding
+        android:bottom="5dp"
+        android:left="5dp"
+        android:right="5dp"
+        android:top="5dp"/>
+    <corners android:radius="50dp"/>
+    <solid android:color="@color/gris_clarito"/>
+</shape>
