+++ YIM/app/src/main/res/layout/activity_crear_ejercicios.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".vista.vistas.CrearEjercicios">
+    tools:context=".vista.vista.CrearEjercicios">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
