+++ YIM/app/src/main/res/layout/activity_ejercicios_diarios.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".vista.vistas.EjerciciosDiarios">
+    tools:context=".vista.vista.EjerciciosDiarios">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
