+++ YIM/app/src/main/res/layout/activity_ejercicios_rutinas.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".vista.vistas.EjerciciosRutinas">
+    tools:context=".vista.vista.EjerciciosRutinas">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
