+++ YIM/app/src/main/res/layout/activity_preguntas_registro.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".vista.vistas.PreguntasRegistro">
+    tools:context=".vista.vista.PreguntasRegistro">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
