+++ YIM/app/src/main/res/layout/activity_rutina_semanal.xml
@@ -4,4 +4,4 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".vista.vistas.RutinaSemanal"/>
+    tools:context=".vista.vista.RutinaSemanal"/>
